Implement a framework for dynamic RADIUS server settings
Using a custom authentication form and view, the RADIUS server to use can be
determined based on the request (i.e. the HTTP_HOST header). The assumption is
that different RADIUS servers might be used depending on who is using the site.

Static RADIUS server configuration can still be used by specifying the details
in settings.py and adding RADIUSAuthenticationBackend to the
AUTHENTICATION_BACKENDS setting.